Get pool name from root, use prefix for kernels
Grab the pool name from The root file system. This is all
that zedenv currently does to get the pool.

https://github.com/johnramsden/zedenv/blob/f1721be772252d77cc5269a0e1711c95610b4d1e/zedenv/lib/be.py#L138

In sticking with the existing format previously-used in
systemd-boot plug-in, prefix the boot environments datasets
with 'env/zedenv-'. Just makes it easier to tell what is
a boot environment we have created, and something that was
prior to zedenv.

We also might as well remove the btrfs check if this is going
to be zedenv specific.